refactor: Create useLogout custom hook
