+++ client/src/Components/index.js
@@ -10,7 +10,6 @@ import PasswordInput from "./Input/PasswordInput/Passwordinput"
 import EmailInputWithValidation from "./Input/EmailInputWithValidation/EmailInputWithValidation"
 import UserNameInputWithValidation from "./Input/UserNameInputWithValidation/UserNameInputWithValidation"
 import Loading from "./LoadingComponent/Loading"
-import ArticleForm from "./ArticleForm/ArticleForm"
 import MarkdownEditor from "./MarkdownEditor/MarkdownEditor"
 
 
@@ -27,6 +26,5 @@ export {
     UserNameInputWithValidation,
     PasswordInput,
     Loading,
-    ArticleForm,
     MarkdownEditor
 }
