+++ client/src/Hooks/index.js
@@ -1,0 +1,7 @@
+import useAuth from "./useAuth"
+import useLogout from "./useLogout"
+
+export{
+    useAuth,
+    useLogout,
+}
