+++ client/src/Pages/CreateArticle/CreateArticle.js
@@ -4,7 +4,8 @@ import './CreateArticle.css'
 
 import { RiSave2Line } from 'react-icons/ri'
 
-import { ArticleForm, MarkdownEditor, Modal } from 'Components'
+import { MarkdownEditor, Modal } from 'Components'
+import ArticleForm from './ArticleForm/ArticleForm'
 
 export default function CreateArticle() {
     const [markdown, setMarkdown] = useState('')
